+++ visual_graph_datasets/util.py
@@ -1,54 +1,63 @@
-import json
-import os
-import sys
-import pathlib
-import logging
 import difflib
-import subprocess
 import importlib.util
+import json
+import logging
+import os
+import pathlib
+import subprocess
+import sys
 import typing as t
 from collections import defaultdict
+from typing import Dict, List
 
 import click
 import jinja2 as j2
-import numpy as np
 import matplotlib.colors as mcolors
+import numpy as np
 
-from visual_graph_datasets.config import HOME_PATH, FOLDER_PATH, CONFIG_PATH
 from visual_graph_datasets.config import Config
 
 PATH = pathlib.Path(__file__).parent.absolute()
 EXPERIMENTS_PATH = os.path.join(PATH, 'experiments')
-TEMPLATES_PATH = os.path.join(PATH, 'templates')
-TEMPLATE_ENV = j2.Environment(
-    loader=j2.FileSystemLoader(TEMPLATES_PATH)
-)
-TEMPLATE_ENV.globals.update({
-    'os': os
-})
-
 VERSION_PATH = os.path.join(PATH, 'VERSION')
 
+# --- jinja2 templates ---
+TEMPLATES_PATH = os.path.join(PATH, 'templates')
+TEMPLATE_ENV = j2.Environment(
+    loader=j2.FileSystemLoader(TEMPLATES_PATH),
+)
+TEMPLATE_ENV.globals.update(
+    {
+        'os': os,
+    },
+)
+
+# --- logging ---
+# The null logger can serve as the default parameter for any function that
+# requires a logger.
 NULL_LOGGER = logging.Logger('null')
 NULL_LOGGER.addHandler(logging.NullHandler())
 
 # 24.12.24
-# Some functions will required a "channel_infos" parameter which is supposed to be a dictionary 
-# whose keys are the integer indices of the explanation channels and the corresponding values are 
-# dictionaries themselves which associate string property names for additional information about 
-# the corresponding channel. 
-# This dictionary provides a default value for those cases. The defaultdict implementation ensures 
-# that it works for any channel index given as a key and the it will return always the same 
+# Some functions will required a "channel_infos" parameter which is supposed to be a dictionary
+# whose keys are the integer indices of the explanation channels and the corresponding values are
+# dictionaries themselves which associate string property names for additional information about
+# the corresponding channel.
+# This dictionary provides a default value for those cases. The defaultdict implementation ensures
+# that it works for any channel index given as a key and the it will return always the same
 # info dict with the default values.
-DEFAULT_CHANNEL_INFOS: dict[str, t.Any] = defaultdict(lambda: {
-    'name': 'channel',
-    'color': 'lightgray',
-})
+DEFAULT_CHANNEL_INFOS: Dict[str, t.Any] = defaultdict(
+    lambda: {
+        'name': 'channel',
+        'color': 'lightgray',
+    },
+)
 
 
 def get_experiment_path(name: str) -> str:
     """
-    Given an experiment file ``name``, this function returns the absolute path to that experiment module.
+    Given an experiment file ``name``, this function returns the absolute path
+    to that experiment module.
 
     :param name: The file name of the experiment, which is part of this package.
 
@@ -63,7 +72,7 @@ def get_version() -> str:
 
     :return: the version string
     """
-    with open(VERSION_PATH, mode='r') as file:
+    with open(VERSION_PATH) as file:
         content = file.read()
 
     version_string = content.replace(' ', '').replace('\n', '')
@@ -72,8 +81,9 @@ def get_version() -> str:
 
 def dynamic_import(module_path: str, name: str = 'pre_process'):
     """
-    Given an absolute path ``module_path`` to a python module, this function will dynamically import that
-    module and return the module object, which can be used to access the contents of that module.
+    Given an absolute path ``module_path`` to a python module, this function will
+    dynamically import that module and return the module object, which can be used
+    to access the contents of that module.
 
     :param str module_path: The absolute path to the module to be imported
     :param str name: The string name to be assigned to that module
@@ -82,8 +92,9 @@ def dynamic_import(module_path: str, name: str = 'pre_process'):
     """
     spec = importlib.util.spec_from_file_location(name, module_path)
     module = importlib.util.module_from_spec(spec)
-    # 24.03.2023 - I have learned that this is rather important to add this as well because if this line
-    # is missing that will screw a lot of "inspect" shenanigans in the imported module
+    # 24.03.2023 - I have learned that this is rather important to add this as well
+    # because if this line is missing that will screw a lot of "inspect" shenanigans
+    # in the imported module
     sys.modules[name] = module
     spec.loader.exec_module(module)
 
@@ -94,34 +105,42 @@ def get_dataset_path(dataset_name: str, datasets_path=Config().get_datasets_path
     """
     Returns the absolute path to the dataset folder identified by the given ``dataset_name``
 
-    :param dataset_name: The string name of the dataset whose absolute folder path is to be retrieved
+    :param dataset_name: The string name of the dataset whose absolute folder path is to be
+        retrieved
     :return: The absolute folder path of the dataset
     """
     if not os.path.exists(datasets_path):
-        raise FileNotFoundError(f'The datasets root folder does not yet exist. This indicates that no '
-                                f'datasets have been downloaded yet or that the wrong root path is '
-                                f'configured.')
+        raise FileNotFoundError(
+            'The datasets root folder does not yet exist. This indicates that no '
+            'datasets have been downloaded yet or that the wrong root path is '
+            'configured.',
+        )
 
     dataset_path = os.path.join(datasets_path, dataset_name)
 
-    # The primary value of this function is supposed to be that we immediately check if this datasets even
-    # exists and then raise an error instead of that happening at some other point in the user code.
+    # The primary value of this function is supposed to be that we immediately check
+    # if this datasets even exists and then raise an error instead of that happening
+    # at some other point in the user code.
     if not os.path.exists(dataset_path):
-        dataset_names: t.List[str] = os.listdir(datasets_path)
-        similarities = [(name, difflib.SequenceMatcher(None, dataset_name, name).ratio())
-                        for name in dataset_names]
+        dataset_names: List[str] = os.listdir(datasets_path)
+        similarities = [
+            (name, difflib.SequenceMatcher(None, dataset_name, name).ratio())
+            for name in dataset_names
+        ]
         similarities = sorted(similarities, key=lambda tupl: tupl[1], reverse=True)
-        raise FileNotFoundError(f'There is no dataset of the name "{dataset_name}" in the root dataset '
-                                f'folder "{datasets_path}"! '
-                                f'Did you mean: "{similarities[0][0]}"?')
+        raise FileNotFoundError(
+            f'There is no dataset of the name "{dataset_name}" in the root dataset '
+            f'folder "{datasets_path}"! '
+            f'Did you mean: "{similarities[0][0]}"?',
+        )
 
     return dataset_path
 
 
-def merge_optional_lists(*args) -> t.List[t.Any]:
+def merge_optional_lists(*args) -> List[t.Any]:
     """
-    Given a list of arguments, this function will merge all the arguments which are of the type "list" and
-    will simply ignore all other arguments, including None values.
+    Given a list of arguments, this function will merge all the arguments
+    which are of the type "list" and will simply ignore all other arguments, including None values.
 
     :param args:
     :return: the merged list
@@ -138,13 +157,14 @@ def merge_nested_dicts(original: dict, update: dict) -> dict:
     """
     Merges the new values of the ``update`` dict into the ``original`` dict in a nested manner.
 
-    That means that the merge is executed separately for all nested sub dictionary structures with
-    the same key. If a key does not exist in the original dict but in the update, the key is added to the
-    original as is. If the update has a non-dict value where the original has a dict value at the same key,
-    the original version is overwritten with the update version.
+    That means that the merge is executed separately for all nested sub dictionary
+    structures with the same key. If a key does not exist in the original dict but
+    in the update, the key is added to the original as is. If the update has a non-dict
+    value where the original has a dict value at the same key, the original version
+    is overwritten with the update version.
 
-    Note: Does not perform copies! Updating of mutable objects may result in references being shared by the
-    two input dictionaries in the end.
+    Note: Does not perform copies! Updating of mutable objects may result in references
+    being shared by the two input dictionaries in the end.
 
     :param original: The original dict to be modified
     :param update: the update dict containing the new values to be added to the original.
@@ -152,7 +172,6 @@ def merge_nested_dicts(original: dict, update: dict) -> dict:
     :returns: the merged version of the original dict
     """
     for key, value in update.items():
-
         if key in original:
             if isinstance(original[key], dict) and isinstance(update[key], dict):
                 merge_nested_dicts(original[key], update[key])
@@ -165,30 +184,32 @@ def merge_nested_dicts(original: dict, update: dict) -> dict:
     return original
 
 
-def array_normalize(array: np.ndarray
-                    ) -> np.ndarray:
+def array_normalize(
+    array: np.ndarray,
+) -> np.ndarray:
     """
-    Normalizes all the values of the given ``array`` into a range between 0 and 1 based on the values 
-    maximum and minimum values of the array.
-    
+    Normalizes all the values of the given ``array`` into a range between 0 and 1
+    based on the values maximum and minimum values of the array.
+
     :param array: The array to be normalized
-    
-    :returns: The normalized array 
+
+    :returns: The normalized array
     """
     norm = mcolors.Normalize(vmin=np.min(array), vmax=np.max(array))
     return np.vectorize(norm)(array)
 
 
-def binary_threshold(array: np.ndarray,
-                     threshold: float,
-                     ) -> np.ndarray:
+def binary_threshold(
+    array: np.ndarray,
+    threshold: float,
+) -> np.ndarray:
     """
-    Applies a binary threshold to the given ``array``. All values that are greater than the given 
+    Applies a binary threshold to the given ``array``. All values that are greater than the given
     ``threshold`` will be set to 1, all other values will be set to 0.
-    
+
     :param array: The array to be thresholded
     :param threshold: The threshold value
-    
+
     :returns: The thresholded binary array
     """
     binary = np.zeros_like(array)
@@ -196,36 +217,37 @@ def binary_threshold(array: np.ndarray,
     return binary
 
 
-
-def edge_importances_from_node_importances(edge_indices: np.ndarray,
-                                           node_importances: np.ndarray,
-                                           calc_cb: t.Callable = lambda v1, v2: (v1 + v2) / 2,
-                                           ignore_with_zero: bool = True
-                                           ) -> np.ndarray:
+def edge_importances_from_node_importances(
+    edge_indices: np.ndarray,
+    node_importances: np.ndarray,
+    calc_cb: t.Callable = lambda v1, v2: (v1 + v2) / 2,
+    ignore_with_zero: bool = True,
+) -> np.ndarray:
     """
-    This function can be used to calculate an edge importances array based on the ``edge_indices`` of a
-    graph and the ``node_importances``.
+    This function can be used to calculate an edge importances array based on the
+    ``edge_indices`` of a graph and the ``node_importances``.
 
-    This function will iterate over all the edges and compute the edge importance of that edge as a function
-    of the two node importance arrays of the two nodes which form that edge.
-    The specific function of how that is calculated can be given with ``calc_cb``. By default, it is simply
-    the average.
+    This function will iterate over all the edges and compute the edge importance of that edge
+    as a function of the two node importance arrays of the two nodes which form that edge.
+    The specific function of how that is calculated can be given with ``calc_cb``.
+    By default, it is simply the average.
 
-    :param edge_indices: A numpy array with the shape (E, 2) where E is number of edges in the graph. This
-        array should consist of integer tuples, where the integer values are the node indices that define
-        the edge
-    :param node_importances: A numpy array of the shape (V, K) where V is the number of nodes in the graph
-        and K is the number of distinct explanations.
-    :param calc_cb: A callback function which takes two arguments (both array of shape (K, ) ) and should
-        aggregate them into a single array of edge importances of the same shape
-    :param ignore_with_zero: If this flag is True, then every edge importance will automatically be set
-        to 0 if either one of the contributing node importances is 0, regardless of the outcome of the
-        computation. If it is False, the outcome of the computation will be used in those cases.
+    :param edge_indices: A numpy array with the shape (E, 2) where E is number of edges in
+        the graph. This array should consist of integer tuples, where the integer values
+        are the node indices that define the edge
+    :param node_importances: A numpy array of the shape (V, K) where V is the number of nodes
+        in the graph and K is the number of distinct explanations.
+    :param calc_cb: A callback function which takes two arguments (both array of shape (K, ) )
+        and should aggregate them into a single array of edge importances of the same shape
+    :param ignore_with_zero: If this flag is True, then every edge importance will
+        automatically be set to 0 if either one of the contributing node importances is 0,
+        regardless of the outcome of the computation. If it is False, the outcome of
+        the computation will be used in those cases.
 
     :return: A numpy array with the shape (E, K) for the edge importances.
     """
     edge_importances = []
-    for e, (i, j) in enumerate(edge_indices):
+    for _, (i, j) in enumerate(edge_indices):
         node_importance_i = node_importances[i]
         node_importance_j = node_importances[j]
 
@@ -246,8 +268,8 @@ def ensure_folder(path: str) -> None:
     # If the path exists then that's nice and we don't need to do anything at all
     if os.path.exists(path):
         return
-    # This is the base case of the recursion: The immediate parent folder exists but the given path does
-    # not, which means to fix this we can simply create a new folder
+    # This is the base case of the recursion: The immediate parent folder exists but the
+    # given path does not, which means to fix this we can simply create a new folder
     elif not os.path.exists(path) and os.path.exists(parent_path):
         os.mkdir(path)
     # Otherwise more of the nested structure does not exist yet and we enter the recursion
@@ -258,7 +280,6 @@ def ensure_folder(path: str) -> None:
 
 # https://stackoverflow.com/questions/434597
 def open_editor(path: str, config=Config()):
-
     platform_id = config.get_platform()
     if platform_id == 'Darwin':
         subprocess.run(['open', path], check=True)
@@ -274,8 +295,8 @@ def sanitize_input(string: str):
 
 def sanitize_indents(string: str) -> str:
     """
-    Given an input ``string`` with multiple lines, this function will remove all the additional indents from
-    that string.
+    Given an input ``string`` with multiple lines, this function will
+    remove all the additional indents from that string.
 
     Written by ChatGPT.
 
@@ -300,8 +321,7 @@ def sanitize_indents(string: str) -> str:
 
 
 class Batched:
-
-    def __init__(self, data: t.List, batch_size: int):
+    def __init__(self, data: list, batch_size: int):
         self.data = data
         self.batch_size = batch_size
 
@@ -317,12 +337,13 @@ class Batched:
         if num == 0:
             raise StopIteration
 
-        data = self.data[self.current_index:self.current_index+num]
+        data = self.data[self.current_index : self.current_index + num]
         self.current_index += num
         return data
 
 
 # == CUSTOM JINJA FILTERS ==
+
 
 def j2_filter_bold(value: str):
     return click.style(value, bold=True)
@@ -340,35 +361,34 @@ TEMPLATE_ENV.filters['fg'] = j2_filter_fg
 # The following sections deals with additional implementations for the "click" command line library
 # which is used to implement the CLI in this project
 
-class JsonListParamType(click.ParamType):
 
+class JsonListParamType(click.ParamType):
     name = 'json_list'
 
     def convert(self, value, param, ctx):
         if isinstance(value, list):
             return list
 
-        # Given the value we will simply attempt to json decode it and if it does not work we know that
-        # it is not a valid json string.
+        # Given the value we will simply attempt to json decode it and if it does not work we
+        # know that it is not a valid json string.
         try:
             loaded = json.loads(value)
 
-            # Even if the conversion is successful, there is still the option that the actual content of
-            # the string is not actually a list.
+            # Even if the conversion is successful, there is still the option that the actual
+            # content of the string is not actually a list.
             if not isinstance(loaded, list):
-                self.fail(f'The content of the given JSON string does not represent a list!')
+                self.fail('The content of the given JSON string does not represent a list!')
 
             return loaded
 
         except Exception:
-            self.fail(f'The given value is not a valid JSON string!')
+            self.fail('The given value is not a valid JSON string!')
 
 
 JSON_LIST = JsonListParamType()
 
 
 class JsonArrayParamType(JsonListParamType):
-
     name = 'json_array'
 
     def convert(self, value, param, ctx):
@@ -377,7 +397,7 @@ class JsonArrayParamType(JsonListParamType):
 
         # The method of the parent class will completely deal with loading the json string as a list
         # already.
-        loaded = super(JsonArrayParamType, self).convert(value, param, ctx)
+        loaded = super().convert(value, param, ctx)
 
         # Now we just need to try and convert it into a numpy array
         try:
@@ -385,15 +405,17 @@ class JsonArrayParamType(JsonListParamType):
 
             return array
 
-        except Exception as e:
-            self.fail(f'The given list cannot be loaded as a numpy array due to the following error: {exc}')
+        except Exception as exc:
+            self.fail(
+                f'The given list cannot be loaded as a numpy array '
+                f'due to the following error: {exc}',
+            )
 
 
 JSON_ARRAY = JsonArrayParamType()
 
 
 class JsonDictParamType(click.ParamType):
-
     name = 'json_dict'
 
     def convert(self, value, param, ctx):
@@ -403,15 +425,15 @@ class JsonDictParamType(click.ParamType):
         try:
             loaded = json.loads(value)
 
-            # Even if the conversion is successful, there is still the option that the actual content of
-            # the string is not actually a dict.
+            # Even if the conversion is successful, there is still the option that the
+            # actual content of the string is not actually a dict.
             if not isinstance(loaded, dict):
-                self.fail(f'The content of the given JSON string does not represent a dictionary!')
+                self.fail('The content of the given JSON string does not represent a dictionary!')
 
             return loaded
 
         except Exception:
-            self.fail(f'The given value is not a valid JSON string!')
+            self.fail('The given value is not a valid JSON string!')
 
 
 JSON_DICT = JsonDictParamType